refactor: moved logic in to CoreNode
