+++ src/render-drivers/main/MainOnlyNode.ts
@@ -386,25 +386,6 @@ export class MainOnlyNode extends EventEmitter implements INode {
   }
 
   private onTextureLoaded: TextureLoadedEventHandler = (target, dimensions) => {
-    if (this.src) {
-      if (dimensions.width && dimensions.height) {
-        if (isNaN(this.width) && isNaN(this.height)) {
-          this.width = dimensions.width;
-          this.height = dimensions.height;
-        } else {
-          const imageAspectRatio = getImageAspectRatio(
-            dimensions.width,
-            dimensions.height,
-          );
-
-          if (isNaN(this.width)) {
-            this.width = this.height * imageAspectRatio;
-          } else if (isNaN(this.height)) {
-            this.height = this.width / imageAspectRatio;
-          }
-        }
-      }
-    }
     this.emit('txLoaded', dimensions);
   };
